refactor: Extract cases to eventually be autogenerated to a module
